Standardise pass entrypoints
Each pass only introduces one `pub(crate)` function, called `pass`,
which accepts the result of its predecessor as input, and transforms
it to its output type.

Using free functions in the standard interface also reduces what needs
to be imported (just the module, not specific types), and it reduces
the requirements to annotate lifetimes in that public interface,
because they can be more easily inferred.